+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 ### Fixed
 - Remove the tree shade indicator from the comfort tag list ([#356](https://gitlab.heigit.org/climate-action/plugins/walkability/-/work_items/356))
+- Update mobility-tools to v2.0.3 to fix slope for ohsome v1 and v2
 - Define a fixed scale for the detour factor chart and change the histogram to a bar chart ([#365](https://gitlab.heigit.org/climate-action/plugins/walkability/-/work_items/365))
 
 ### Changed